♻️ refactor(frontend): make populateSelectOptions an importable function to reduce redundancy
