+++ src/main/resources/static/js/modules/courses/courses.js
@@ -14,6 +14,8 @@
  * @author Jason
  */
 
+import { loadTableData } from '../../shared/utils/tables/index.js'
+
 import {
 	showToast,
 	toggleButtonLoading,
@@ -23,8 +25,6 @@ import {
 	setupBootstrapSelectDropdownStyles,
 	initializeTooltips,
 } from '../../shared/utils/ui/index.js'
-
-import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import { isValidText } from '../../shared/utils/validators/index.js'
