+++ src/main/resources/static/js/modules/publishers/publishers.js
@@ -14,19 +14,23 @@
  * @author Jason
  */
 
+import { loadTableData } from '../../shared/utils/tables/index.js'
+
+import {
+	loadSelectOptions,
+	populateSelect,
+} from '../../shared/utils/forms/select-options.js'
+
 import {
 	showToast,
 	toggleButtonLoading,
 	toggleModalLoading,
-	populateSelect,
 	placeholderColorSelect,
 	placeholderColorEditSelect,
 	initializeCropper,
 	setupBootstrapSelectDropdownStyles,
 	initializeTooltips,
 } from '../../shared/utils/ui/index.js'
-
-import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
 	isValidText,
@@ -45,45 +49,6 @@ let literaryGenreList = []
 
 // Global variable to handle photo deletion in edit modal
 let deletePhotoFlag = false
-
-async function populateSelectOptions() {
-	try {
-		const response = await fetch('./api/publishers/select-options', {
-			method: 'GET',
-			headers: {
-				Accept: 'application/json',
-			},
-		})
-
-		if (response.status === 204) {
-			console.warn('No data found for select options.')
-			return
-		}
-
-		if (!response.ok) {
-			throw response
-		}
-
-		const data = await response.json()
-
-		nationalityList = data.nationalities
-		literaryGenreList = data.literaryGenres
-	} catch (error) {
-		if (error instanceof Response) {
-			try {
-				const errData = await error.json()
-				console.error(
-					`Error fetching select options (${errData.errorType} - ${error.status}):`,
-					errData.message,
-				)
-			} catch {
-				console.error('Unexpected error:', error.status, await error.text())
-			}
-		} else {
-			console.error('Unexpected error:', error)
-		}
-	}
-}
 
 /*****************************************
  * TABLE HANDLING
@@ -1180,7 +1145,13 @@ $(document).ready(function () {
 	handleAddPublisherForm()
 	handleEditPublisherForm()
 	loadModalData()
-	populateSelectOptions()
+	loadSelectOptions({
+		url: './api/publishers/select-options',
+		onSuccess: (data) => {
+			nationalityList = data.nationalities
+			literaryGenreList = data.literaryGenres
+		},
+	})
 	$('.selectpicker').selectpicker()
 	setupBootstrapSelectDropdownStyles()
 	placeholderColorSelect()
