+++ src/main/resources/static/js/modules/students/students.js
@@ -14,11 +14,17 @@
  * @author Jason
  */
 
+import { loadTableData } from '../../shared/utils/tables/index.js'
+
+import {
+	loadSelectOptions,
+	populateSelect,
+} from '../../shared/utils/forms/select-options.js'
+
 import {
 	showToast,
 	toggleButtonLoading,
 	toggleModalLoading,
-	populateSelect,
 	placeholderColorSelect,
 	placeholderColorEditSelect,
 	placeholderColorDateInput,
@@ -26,8 +32,6 @@ import {
 	initializeTooltips,
 	getCurrentPeruDate,
 } from '../../shared/utils/ui/index.js'
-
-import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
 	isValidDNI,
@@ -44,44 +48,6 @@ import {
 
 // Global list of faculties for the selectpickers
 let facultyList = []
-
-async function populateSelectOptions() {
-	try {
-		const response = await fetch('./api/students/select-options', {
-			method: 'GET',
-			headers: {
-				Accept: 'application/json',
-			},
-		})
-
-		if (response.status === 204) {
-			console.warn('No data found for select options.')
-			return
-		}
-
-		if (!response.ok) {
-			throw response
-		}
-
-		const data = await response.json()
-
-		facultyList = data.faculties
-	} catch (error) {
-		if (error instanceof Response) {
-			try {
-				const errData = await error.json()
-				console.error(
-					`Error fetching select options (${errData.errorType} - ${error.status}):`,
-					errData.message,
-				)
-			} catch {
-				console.error('Unexpected error:', error.status, await error.text())
-			}
-		} else {
-			console.error('Unexpected error:', error)
-		}
-	}
-}
 
 /*****************************************
  * TABLE HANDLING
@@ -1057,7 +1023,12 @@ $(document).ready(function () {
 	handleAddStudentForm()
 	handleEditStudentForm()
 	loadModalData()
-	populateSelectOptions()
+	loadSelectOptions({
+		url: './api/students/select-options',
+		onSuccess: (data) => {
+			facultyList = data.faculties
+		},
+	})
 	$('.selectpicker').selectpicker()
 	setupBootstrapSelectDropdownStyles()
 	placeholderColorSelect()
